+++ web/fill/issues.js
@@ -1,0 +1,144 @@
+const policyData = [
+    {
+        category: "Foreign Policy",
+        category_description: "Foreign Policy Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum ac diam sit amet quam vehicula elementum sed sit amet dui. Curabitur aliquet quam id dui posuere blandit. Curabitur non nulla sit amet nisl tempus convallis quis ac lectus. Proin eget tortor risus. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae; Donec velit neque, auctor sit amet aliquam vel, ullamcorper sit amet ligula. Pellentesque in ipsum id orci porta dapibus. Nulla porttitor accumsan tincidunt.",
+        icon: "foreign_policy.png",
+        issues: [
+            {
+                issue_name: "Military Spending",
+                issue_description: "Debates over defense budget and military investments."
+            },
+            {
+                issue_name: "International Alliances",
+                issue_description: "Discussions on NATO, UN, and other global partnerships."
+            }
+        ]
+    },
+    {
+        category: 'Social Issues',
+        category_description: 'Policies related to civil rights, LGBTQ+ rights, and religious freedoms.',
+        icon: 'social_issues.png',
+        issues: [
+            {
+                issue_name: 'Civil Rights',
+                issue_description: 'Legislation and protections for marginalized groups.'
+            },
+            {
+                issue_name: 'LGBTQ+ Rights',
+                issue_description: 'Policies supporting LGBTQ+ equality and protections.'
+            },
+        ]
+    },
+    {
+        category: 'Immigration',
+        category_description: 'Policies on border security and citizenship.',
+        icon: 'immigration.png',
+        issues: [
+            {
+                issue_name: 'Border Wall',
+                issue_description: 'Debates over building a wall along the US-Mexico border.'
+            },
+            {
+                issue_name: 'Path to Citizenship',
+                issue_description: 'Proposals for legalizing undocumented people'
+            },
+        ]
+    },
+    {
+        category: "Economy",
+        category_description: "Issues regarding taxation, employment, and business.",
+        icon: "economy.png",
+        issues: [
+            {
+                issue_name: "Minimum Wage",
+                issue_description: "Debates over raising or lowering the federal minimum wage."
+            },
+            {
+                issue_name: "Tax Reform",
+                issue_description: "Discussions on restructuring tax laws and rates."
+            },
+            {
+                issue_name: "Trade Policies",
+                issue_description: "Regulations regarding international trade and tariffs."
+            }
+        ]
+    },
+    {
+        category: "Healthcare",
+        category_description: "Policies related to medical care and insurance.",
+        icon: "healthcare.png",
+        issues: [
+            {
+                issue_name: "Medicare Expansion",
+                issue_description: "Debates over expanding or reducing Medicare coverage."
+            },
+            {
+                issue_name: "Prescription Drug Prices",
+                issue_description: "Policies aimed at controlling medication costs."
+            },
+            {
+                issue_name: "Mental Health Services",
+                issue_description: "Efforts to increase access to mental health care."
+            }
+        ]
+    },
+    {
+        category: "Education",
+        category_description: "Funding and reforms for public and private education.",
+        icon: "education.png",
+        issues: [
+            {
+                issue_name: "Student Loan Forgiveness",
+                issue_description: "Proposals to reduce or eliminate student loan debt."
+            },
+            {
+                issue_name: "Public School Funding",
+                issue_description: "Discussions on the allocation of state and federal funds for education."
+            },
+            {
+                issue_name: "Charter Schools",
+                issue_description: "Debates on the role and funding of charter schools."
+            }
+        ]
+    },
+    {
+        category: "Environment",
+        category_description: "Policies addressing climate change and sustainability.",
+        icon: "environment.png",
+        issues: [
+            {
+                issue_name: "Renewable Energy",
+                issue_description: "Investments and regulations for wind, solar, and hydro energy."
+            },
+            {
+                issue_name: "Carbon Emissions",
+                issue_description: "Legislation to reduce carbon footprints and pollution."
+            },
+            {
+                issue_name: "Deforestation",
+                issue_description: "Efforts to curb illegal logging and protect forests."
+            }
+        ]
+    },
+    {
+        category: "Gun Control",
+        category_description: "Regulations regarding firearm ownership and use.",
+        icon: "gun_control.png",
+        issues: [
+            {
+                issue_name: "Background Checks",
+                issue_description: "Expanding or restricting background checks for gun purchases."
+            },
+            {
+                issue_name: "Assault Weapons Ban",
+                issue_description: "Debates over banning high-capacity firearms."
+            },
+            {
+                issue_name: "Gun Ownership",
+                issue_description: "Second Amendment rights and legal challenges."
+            }
+        ]
+    }
+];
+
+export default policyData;
